Removed unnecessary cast from Record init with Builder, fixed enum @Unknown access bug

+++ restlicsharpdata/restlidataintegration/com/linkedin/restli/datagenerator/integration/RecordInUnionTest.cs
@@ -32,7 +32,7 @@ namespace com.linkedin.restli.datagenerator.integration
       // Retrieve data for a
       if (builder.a != null)
       {
-        a = (int)builder.a;
+        a = builder.a;
 
       }
       else
